requirements.txt
langchain
langchain_google_genai
langchain-openai
python-dotenv
pydantic
duckduckgo-search
wikipedia
langchain-community
langchain-openai
openai
langgraph
azure-ai-inference
# Download from GitHub
git clone https://github.com/github/codeql.git
cd codeql
# Build (or download prebuilt from releases)
./scripts/install-dependencies.sh
